Add auto link feature and get_container_name.
When a Selenium Grid Node is started, it will try to automatically link
with a running Grid Hub, and Proxy Server.

In case --skip-hub, or/and --skip-proxy, the link will not be performed.

For performing a link with another running container, you can optionally
specify --link-hub=<container_name> and/or
--link-proxy=<container_name>.